+++ engine/src/camera.h
@@ -10,7 +10,7 @@ namespace engine
 	const float CAMERA_MOUSE_SENSITIVITY = 0.002f;
 	const float CAMERA_SPEED = 2.0f;
 
-	const float MAX_TILT = 0.95 * M_PI_2;
+	const float MAX_TILT = 0.95f * (float)M_PI_2;
 
 	class Camera
 	{
@@ -34,7 +34,7 @@ namespace engine
 		float getYaw();
 		Vector3 getPosition();
 
-		void update(double elapsedSecs, Vector2 cursorDiff);
+		void update(float elapsedSecs, Vector2 cursorDiff);
 
 	private:
 		GLuint uboId = 0;
